fix(genai-text,#8707,#8664): drop stale Périmètre du run caveat + dedup cycle-id from script names
- cell[0]: remove 8-line "Périmètre du run notebook" block (became false
  post-merge + no longer belongs in pedagogical header); restore `---`
  separator between Prérequis and Objectifs to keep section break clean.
- 3 required pedagogical sections retained verbatim from origin/main:
  ## Pourquoi héberger localement ? · ## Modèles locaux recommandés ·
  ## Installation & Import. Content-loss detector: 0 findings, MD cells
  stable 44/44, normalized chars 42373 → 42927.
- Rename scripts: scripts/exec_c939_cells.py →
  scripts/exec_multi_endpoint_cells.py (and 2 siblings). Cycle id no
  longer hardcoded in tool name; script is generic across multi-endpoint
  re-execution cycles.
- 6 notebook cells (34/41/44/51/54/61) updated to reference new script
  name. The c939_run_results.json data artifact path is kept
  intentionally (cycle-specific data, regenerable).
- .gitignore comment updated to reference new script name.

No re-execution required (ai-01 c.35 ask — pure markdown + script rename).
See #8369 #8664
