+++ .gitignore
@@ -834,7 +834,7 @@ analyze_ratio.py
 # Whisper comparison results (regenerable)
 scripts/whisper_comparison_results.json
 
-# Cycle c.939 run results (regenerable via scripts/exec_c939_cells.py)
+# Multi-endpoint run results (regenerable via scripts/exec_multi_endpoint_cells.py)
 MyIA.AI.Notebooks/GenAI/Texte/c939_run_results.json
 
 # Final model files referenced by executed notebooks (tracked via git LFS)
