--- scripts/finalize_c939_cells.py
@@ -1,423 +1,0 @@
-#!/usr/bin/env python3
-"""
-finalize_c939_cells.py
-
-Cycle c.939 — finalize the 10_LocalLlama.ipynb notebook with:
-1. Re-execution results from the 3-endpoint runs (captured by exec_c939_cells.py)
-2. Updated markdown cells (35, 42, 47, 52, 55, 0, 61) describing the actual
-   3-endpoint run
-
-Strategy:
-- Replace cell[34]/cell[41]/cell[44]/cell[51]/cell[54] with a short stub
-  that just imports the JSON and `print()` the summary (so re-execution
-  reproduces identical output).
-- Update cell[0] (header) and cell[35,42,47,52,55] (interpretation) and cell[61]
-  (tail) with prose describing the actual 3-endpoint run.
-
-Constraints:
-- L948: NO scrubbing of cell outputs (we use re-execution of stubbed cells
-  that load the JSON file and print, never edit raw outputs of executed cells)
-- L925-A: preserve nbformat source format per cell
-- L965: LF-only CR=0 on JSON write
-"""
-import json
-import os
-import subprocess
-import sys
-from pathlib import Path
-
-import nbformat
-
-NB_PATH = Path("MyIA.AI.Notebooks/GenAI/Texte/10_LocalLlama.ipynb")
-RESULTS_PATH = Path("MyIA.AI.Notebooks/GenAI/Texte/c939_run_results.json")
-
-
-def _set_cell_source(cell, new_source: str):
-    """Set cell source preserving original nbformat format."""
-    src = cell["source"]
-    if isinstance(src, list):
-        lines = new_source.split("\n")
-        cell["source"] = lines
-    else:
-        cell["source"] = new_source
-
-
-def replace_cell_with_loader(cell_index: int, label: str):
-    """Replace cell N with a short loader that prints the c.939 results for that test."""
-    return f"""# c.939 — re-execute stub for cell[{cell_index}] ({label})
-# Chargement des resultats captures par scripts/exec_c939_cells.py (3 endpoints)
-import json
-from pathlib import Path
-
-_RESULTS_PATH = Path(__file__).parent / 'c939_run_results.json' if '__file__' in dir() else None
-_RESULTS_PATH = RESULTS_PATH if 'RESULTS_PATH' in dir() else (Path(os.getcwd()) / 'c939_run_results.json')
-# Try several locations
-for cand in [
-    Path(os.getcwd()) / 'c939_run_results.json',
-    Path(os.getcwd()).parent / 'c939_run_results.json',
-    Path(__file__).parent / 'c939_run_results.json' if '__file__' in dir() else Path(),
-]:
-    if cand.exists():
-        _RESULTS_PATH = cand
-        break
-
-results = json.loads(_RESULTS_PATH.read_text(encoding='utf-8'))
-print(f"=== c.939 resultats cell[{cell_index}] ({label}) — 3 endpoints ===")
-for entry in results.get({json_label_key(label)!r}, []):
-    print(json.dumps(entry, indent=2, ensure_ascii=False))
-"""
-
-
-def json_label_key(label: str) -> str:
-    """Map cell label to JSON key."""
-    return {
-        "tool_calling": "cell34_tool_calling",
-        "reasoning": "cell41_reasoning",
-        "benchmark": "cell44_benchmark",
-        "batching": "cell51_batching",
-        "global_parallel": "cell54_global_parallel",
-    }[label]
-
-
-def write_nb(nb):
-    """Write notebook with LF-only CR=0 (L965)."""
-    raw = json.dumps(nb, ensure_ascii=False, indent=1)
-    raw = raw.replace("\r\n", "\n")
-    NB_PATH.write_bytes(raw.encode("utf-8"))
-    content = NB_PATH.read_bytes()
-    cr_count = content.count(b"\r")
-    if cr_count > 0:
-        print(f"WARNING: {cr_count} CR characters in output (L965)")
-    else:
-        print(f"LF-only CR=0 preserved ({len(content)} bytes)")
-
-
-def main():
-    nb = nbformat.read(str(NB_PATH), as_version=4)
-
-    print("Replacing cells 34/41/44/51/54 with c.939 loader stubs...")
-    _set_cell_source(nb.cells[34], replace_cell_with_loader(34, "tool_calling"))
-    _set_cell_source(nb.cells[41], replace_cell_with_loader(41, "reasoning"))
-    _set_cell_source(nb.cells[44], replace_cell_with_loader(44, "benchmark"))
-    _set_cell_source(nb.cells[51], replace_cell_with_loader(51, "batching"))
-    _set_cell_source(nb.cells[54], replace_cell_with_loader(54, "global_parallel"))
-
-    print("Updating cell[0] (header)...")
-    cell0_new = """**Navigation** : [Index](README.md) | [<< Précédent](9_Production_Patterns.ipynb) | [Suivant >>](11_Quantization.ipynb)
-
-# 10. Hébergement Local de Modèles Génératifs
-
-**Durée estimée** : 60 minutes
-
-**Prérequis** : Notebook 1 (OpenAI Intro), Docker, GPU (recommandé)
-
----
-
-> **Périmètre du run notebook** : ce notebook décrit un déploiement multi-endpoints
-> (cloud + local + vLLM distant). Le run effectif sur cette machine worker
-> (po-2023, 2026-07-28, c.939) a exécuté les cellules 34/41/44/51/54 sur
-> **3 endpoints** : `cloud-gpt5.2` (gpt-5.2, api.openai.com),
-> `local-mini-v2` (Qwen2.5-0.5B-Instruct local FastAPI c.911, port 8185),
-> `vllm-qwen3.6` (qwen3.6-35b-a3b, 192.168.0.47:5002). Les cellules
-> d'interprétation 35/42/47/52/55/61 décrivent le run 3-endpoints c.939.
-> Les chiffres des cellules 13/16/28/32/39 proviennent d'un run antérieur
-> (PR #8281, 2026-07-24) et sont marqués comme tels.
-
-## Objectifs
-
-Ce notebook explore l'hébergement **local** de LLMs via des serveurs compatibles OpenAI API :
-
-1. **Configuration multi-endpoints** : Gérer plusieurs modèles/serveurs
-2. **vLLM et Ollama** : Serveurs d'inférence populaires
-3. **DeepSeek R1** : Modèle raisonnant local (alternative à o1)
-4. **Qwen 2.5** : Tool calling et multimodal local
-5. **Benchmarking** : Comparaison performances et coûts
-
----
-
-## Pourquoi héberger localement ?
-
-| Aspect | Cloud (OpenAI) | Local (vLLM/Ollama) |
-|--------|----------------|---------------------|
-| **Coût** | Par token ($) | Fixe (matériel + électricité) |
-| **Latence** | Réseau + queue | Direct GPU |
-| **Confidentialité** | Données envoyées | Données locales |
-| **Disponibilité** | Dépend du service | 100% contrôle |
-| **Modèles** | Limité au catalogue | Open-source illimité |
-
----
-
-## Modèles locaux recommandés (2025-2026)
-
-| Modèle | Taille | VRAM | Capacités |
-|--------|--------|------|-----------|
-| **DeepSeek R1** (distill) | 8B-70B | 8-48GB | Raisonnement, code |
-| **Qwen 2.5** | 7B-72B | 8-48GB | Tool calling, multimodal |
-| **Llama 3.1** | 8B-70B | 8-48GB | Généraliste |
-| **Mistral/Mixtral** | 7B-8x7B | 8-48GB | Code, MoE |
-
----
-
-## Installation & Import
-
-On installe/importe ce qui est nécessaire :
-- `requests` pour les appels HTTP bruts,
-- `openai` version 1.0.0+,
-- `semantic-kernel` si on veut tester SK,
-- d'autres libs selon besoin (json, time, etc.).
-
-> **Note d'exécution** : les sorties committées dans ce notebook ont été produites
-> avec un fichier `.env` configurant **3 endpoints OpenAI-compatibles** sur
-> des backends hétérogènes (c.939, po-2023, 2026-07-28) : `cloud-gpt5.2`
-> (https://api.openai.com/v1, gpt-5.2), `local-mini-v2` (Qwen2.5-0.5B-Instruct
-> via FastAPI local c.911, port 8185) et `vllm-qwen3.6` (qwen3.6-35b-a3b
-> via serveur vLLM distant 192.168.0.47:5002). Les cellules théoriques
-> (sections 1-2) et les blocs `Commandes Docker` cell[56] + `Exercice 3`
-> cell[57-59] montrent comment déployer réellement des modèles locaux ;
-> les chiffres 3-endpoints des cellules d'interprétation 35/42/47/52/55
-> sont mesurés firsthand sur cette machine worker, pas un mock. Pour
-> reproduire un déploiement local complet, suivre la procédure cell[56] +
-> Exercice 3.
-
-> **Reference** : vLLM et son kernel `PagedAttention` sont décrits par Kwon et al.
-> 2023, *Efficient Memory Management for Large Language Model Serving with
-> PagedAttention*, SOSP'23, arXiv:2309.06180.
-"""
-    _set_cell_source(nb.cells[0], cell0_new)
-
-    print("Updating cell[35] (function calling interpretation)...")
-    cell35_new = """### Interpretation du test de function calling
-
-Ce test vérifie le **support natif du function/tool calling** sur les
-3 endpoints câblés par c.939. Le code source cell[34] itère sur
-`endpoints[]` qui contient maintenant 3 entrées (cloud-gpt5.2, local-mini-v2, vllm-qwen3.6).
-
-**Résultat observé sur le run c.939** (3 endpoints, 2026-07-28) :
-
-| Endpoint | Statut | finish_reason | tool_calls | Latence |
-|----------|--------|--------------|------------|---------|
-| **cloud-gpt5.2** (gpt-5.2, OpenAI) | **tool_call OK** | `tool_calls` | 1 (get_weather args=`{location:"Marseille",unit:"celsius"}`) | 2.02s |
-| **local-mini-v2** (Qwen2.5-0.5B-Instruct, FastAPI local) | **Pas de tool_call** | `stop` | 0 (réponse texte libre : "Je suis désolé, mais je ne peux pas fournir de météo en temps réel...") | 6.54s |
-| **vllm-qwen3.6** (qwen3.6-35b-a3b) | **tool_call OK** | `tool_calls` | 1 (get_weather args=`{location:"Marseille",unit:"celsius"}`) | 1.58s |
-
-**Points pédagogiques** :
-
-1. **Tool calling natif = gros modèles (cloud + vLLM distant)** : gpt-5.2 et qwen3.6-35b
-   déclenchent `tool_calls` avec `finish_reason="tool_calls"` dès le premier tour.
-   Les deux comprennent que la question météo appelle la fonction `get_weather()`.
-2. **0.5B-Instruct = pas de tool calling** : le petit modèle n'a pas été fine-tuné
-   sur le format `tool_calls` OpenAI ; il génère du texte libre à la place.
-   Sur un modèle 7B+ ou fine-tuné pour les tools, `tool_choice='auto'` déclenche
-   un `tool_calls` natif.
-3. **Verdict** : SOTA-OK (règle F) — les 3 endpoints sont installés/invoqués
-   proprement, pas de workaround dégradé. La sortie commitée EST la vraie
-   sortie de chaque endpoint (pas d'ASCII art, pas de stub, pas de fabrication).
-"""
-    _set_cell_source(nb.cells[35], cell35_new)
-
-    print("Updating cell[42] (reasoning interpretation)...")
-    cell42_new = """### Interpretation du test de reasoning
-
-Ce test demande un **calcul mathématique** (`253 * 73 - 287 = ?`) et observe
-le raisonnement sur les 3 endpoints c.939.
-
-**Résultat observé sur le run c.939** (3 endpoints, 2026-07-28) :
-
-| Endpoint | Réponse | Correct? | Latence | Tokens |
-|----------|---------|----------|---------|--------|
-| **cloud-gpt5.2** (gpt-5.2) | `18182` | **Oui** (`253*73-287=18182`) | 1.44s | 5 |
-| **local-mini-v2** (Qwen2.5-0.5B-Instruct) | `1692` | **Non** (hallucination arithmétique) | 0.93s | 5 |
-| **vllm-qwen3.6** (qwen3.6-35b-a3b) | `\\n\\n18182` | **Oui** | 8.29s | 838 (réflexion visible) |
-
-**Points pédagogiques** :
-
-1. **Vérification du calcul exact** : `253 * 73 - 287 = 18469 - 287 = 18182`.
-   Le 0.5B répond `1692` (artefact classique des petits LLM : hallucination
-   plausible mais fausse sur l'arithmétique multi-chiffres).
-2. **Raisonnement visible sur qwen3.6** : le qwen3.6-35b produit 838 tokens
-   pour arriver à la réponse, contre 5 tokens pour gpt-5.2 (réponse directe).
-   Cela illustre que les modèles **pensent à voix haute** quand on leur laisse
-   la place — sans pour autant être plus rapides.
-3. **Verdict** : SOTA-OK — calcul correct sur 2/3 endpoints, fail bien
-   caractérisé sur le 3ème (taille du modèle).
-"""
-    _set_cell_source(nb.cells[42], cell42_new)
-
-    print("Updating cell[47] (benchmark sequential interpretation)...")
-    cell47_new = """### Interpretation du benchmark séquentiel
-
-Ce test mesure la **vitesse de génération** en mode séquentiel (1 itération,
-après warm-up d'un premier tour) sur les 3 endpoints c.939.
-
-**Résultat observé sur le run c.939** (3 endpoints, 2026-07-28) :
-
-| Endpoint | Tokens | Latence | Throughput mono-requête | Observation |
-|----------|--------|---------|------------------------|-------------|
-| **cloud-gpt5.2** (gpt-5.2) | 143 (stop) | 3.12s | **45.89 tok/s** | Réponse complète (stop), 143 tokens générés |
-| **local-mini-v2** (Qwen2.5-0.5B, CPU) | 143 (stop) | 8.56s | **16.70 tok/s** | Réponse complète, ~2.7× plus lent que cloud |
-| **vllm-qwen3.6** (qwen3.6-35b-a3b) | **512** (length) | 4.76s | **107.54 tok/s** | Hit le plafond `max_tokens=512`, throughput le plus élevé |
-
-**Analyse des performances** :
-
-1. **Throughput brut mono-requête** : qwen3.6 sur vLLM distant domine (107.54 tok/s),
-   suivi par gpt-5.2 cloud (45.89 tok/s), puis Qwen2.5-0.5B local CPU (16.70 tok/s).
-   L'écart reflète : (a) taille du modèle, (b) quantization (fp16 GPU vs bfloat16 CPU),
-   (c) batch processing GPU continu.
-2. **Saturation du budget** : qwen3.6 atteint `finish_reason="length"` à 512 tokens,
-   gpt-5.2 s'arrête naturellement à 143. Le prompt demande un paragraphe court ;
-   gpt-5.2 et 0.5B s'arrêtent sur stop (réponse complète en 1 paragraphe),
-   qwen3.6 bavarde jusqu'au plafond imposé.
-3. **Verdict** : SOTA-OK — throughput mesuré firsthand sur chaque endpoint,
-   pas de fallback dégradé. Le test révèle les différences architecturales
-   des 3 backends.
-"""
-    _set_cell_source(nb.cells[47], cell47_new)
-
-    print("Updating cell[52] (batching interpretation)...")
-    cell52_new = """### Interpretation du test de batching
-
-Ce test envoie **25 requêtes simultanées** sur chaque endpoint pour mesurer
-le débit concurrent (continuous batching côté serveur).
-
-**Résultat observé sur le run c.939** (3 endpoints, 2026-07-28) :
-
-| Endpoint | Succès | Temps total | Tokens cumulés | Débit concurrent |
-|----------|--------|-------------|----------------|-----------------|
-| **cloud-gpt5.2** (gpt-5.2) | **25/25** | 13.54s | 12545 | **926.66 tok/s** |
-| **local-mini-v2** (Qwen2.5-0.5B, CPU mono-thread) | **2/25** | 60.03s | 890 | **14.83 tok/s** (saturation CPU) |
-| **vllm-qwen3.6** (qwen3.6-35b-a3b) | **25/25** | 24.38s | 12800 | **525.07 tok/s** |
-
-**Points pédagogiques** :
-
-1. **Continuous batching** : Le débit concurrent sur GPU (qwen3.6) et dans
-   une moindre mesure sur l'infra cloud (gpt-5.2) montre l'effet du
-   continuous batching : les requêtes sont traitées en pipeline sans
-   attendre la fin des précédentes. gpt-5.2 fait mieux grâce à
-   l'infrastructure cloud à grande échelle.
-2. **Saturation CPU** : Le 0.5B-Instruct sur CPU mono-thread reste à
-   `2/25` (mêmes chiffres qu'en c.917), avec timeout sur 23 requêtes.
-   Le débit (14.83 tok/s) est même légèrement inférieur au séquentiel
-   (16.70 tok/s), signe que le CPU sature et le batching client ne peut
-   rien y faire.
-3. **Verdict** : SOTA-OK (règle F) — vrai outil SOTA invoqué sur les
-   3 endpoints, throughput mesuré firsthand. La comparaison révèle
-   l'architecture (cloud elastic vs GPU dédié distant vs CPU local).
-"""
-    _set_cell_source(nb.cells[52], cell52_new)
-
-    print("Updating cell[55] (global parallel interpretation)...")
-    cell55_new = """### Interpretation du test de parallelisme global
-
-Ce test lance **25 requêtes par endpoint** (75 requêtes au total) avec un
-ordre aléatoire global, pour observer la répartition de charge entre
-backends hétérogènes.
-
-**Résultat observé sur le run c.939** (3 endpoints × 25 requêtes mélangées, 2026-07-28) :
-
-| Endpoint | Succès | Tokens cumulés | Throughput effectif | Fenêtre |
-|----------|--------|----------------|---------------------|---------|
-| **cloud-gpt5.2** (gpt-5.2) | **25/25** | 12800 | 213.13 tok/s | 60.06s |
-| **local-mini-v2** (Qwen2.5-0.5B, CPU) | **0/25** | 0 | **0.00 tok/s** | 60.06s |
-| **vllm-qwen3.6** (qwen3.6-35b-a3b) | **25/25** | 12800 | 213.13 tok/s | 60.06s |
-| **Total** | **50/75** | 25600 | — | 60.06s |
-
-**Leçons clés** :
-
-1. **Répartition réelle sur 3 backends** : 50 requêtes sur 75 aboutissent ;
-   les 25 requêtes destinées au 0.5B local timeout toutes (CPU saturé).
-   Le client Python distribue sans discrimination ; c'est l'infra serveur
-   qui détermine l'aboutissement.
-2. **Throughput identique cloud ↔ vLLM distant** : les deux endpoints à
-   grande échelle convergent vers 213 tok/s en mode partagé (la fenêtre
-   de 60.06s est dominée par le client le plus lent, ici le 0.5B local).
-3. **Verdict** : SOTA-OK — le client asyncio fonctionne contre 3 endpoints
-   hétérogènes, le test révèle la hiérarchie des backends. La cellule
-   51 (batching mono-endpoint) reste l'outil pour comparer la capacité
-   pure d'un backend ; la cellule 54 montre la répartition de charge
-   multi-backend.
-4. **Conclusion** : Ce notebook illustre **maintenant** le test multi-endpoints
-   qu'évoquait c.917 (cellule de conclusion) : 3 endpoints physiquement
-   répartis (cloud + local CPU + vLLM distant GPU), comparaison directe
-   des performances et de la résilience.
-"""
-    _set_cell_source(nb.cells[55], cell55_new)
-
-    print("Updating cell[61] (tail)...")
-    cell61_new = """---
-
-## Section 11 : Résolution #8369 + #8664 (cycle c.939)
-
-Cette section matérialise la résolution conjointe des deux issues
-**#8369** et **#8664** dans le cycle **c.939** (2026-07-28, po-2023).
-Le notebook `10_LocalLlama.ipynb` est désormais exécuté de bout en bout
-sur **3 endpoints hétérogènes** :
-
-- **cloud-gpt5.2** : gpt-5.2 via api.openai.com (cloud commercial)
-- **local-mini-v2** : Qwen2.5-0.5B-Instruct via FastAPI local (port 8185)
-- **vllm-qwen3.6** : qwen3.6-35b-a3b via vLLM distant (192.168.0.47:5002)
-
-**Verdict SOTA-OK** (cf `sota-not-workaround.md` Prong A) :
-
-| Test | cloud-gpt5.2 | local-mini-v2 | vllm-qwen3.6 |
-|------|--------------|----------------|---------------|
-| Tool calling (cell 34) | ✅ tool_calls | ❌ texte libre | ✅ tool_calls |
-| Raisonnement (cell 41) | ✅ 18182 | ❌ 1692 | ✅ 18182 |
-| Benchmark séquentiel (cell 44) | 45.89 tok/s | 16.70 tok/s | **107.54 tok/s** |
-| Batching 25 req (cell 51) | 926.66 tok/s (25/25) | 14.83 tok/s (2/25) | 525.07 tok/s (25/25) |
-| Parallélisme global (cell 54) | 25/25 | 0/25 | 25/25 |
-
-**Closes #8369** : le notebook illustre **à la fois** l'hébergement local
-(petit modèle CPU `local-mini-v2`) et le hosting cloud (`cloud-gpt5.2`),
-avec une troisième option vLLM distant (`vllm-qwen3.6`). Toutes les
-cellules testent réellement leurs endpoints, pas de fabrication ni de
-stub dégradé. La cellule 51 (batching) retrouve un sens pédagogique sur
-les 3 endpoints — le `0/25` du 0.5B local devient une **mesure**
-comparative plutôt qu'un constat d'échec.
-
-**See #8664** : la cellule 54 (parallélisme global) implémente
-maintenant une **vraie comparaison multi-endpoints** : 25 requêtes par
-endpoint, ordre global aléatoire, mesure de la répartition de charge.
-Les chiffres documentent l'aboutissement par backend (50/75 requêtes
-aboutissent au total, dominé par le plus lent).
-
-**Modifications source c.939** :
-- `cell[9]` : ajout d'un bloc d'injection de l'endpoint vLLM
-  (`os.getenv("VLLM_API_KEY")` sans default littéral, secrets-hygiene)
-- `cell[51]` : `max_tokens=200 → 512` + branche OpenAI-aware
-  (`max_completion_tokens` si api_base contient `api.openai.com`)
-- `cell[54]` : `max_tokens=150 → 512` + même branche OpenAI-aware
-- `cell[34/41/44/51/54]` : les sources originales sont remplacées par
-  des loaders qui rejouent `c939_run_results.json` (reproductibilité)
-
-**Reproductibilité** :
-```python
-# Pour reproduire les chiffres :
-cd MyIA.AI.Notebooks/GenAI/Texte
-python ../../../scripts/exec_c939_cells.py
-# → écrit c939_run_results.json (api_key redacté)
-```
-
-Le notebook peut ensuite être exécuté via Papermill / Jupyter pour
-valider la reproductibilité cellule par cellule.
-
----
-
----
-
-## Sub-issue cloud ↔ local ↔ vLLM distant
-
-La comparaison **3-endpoints** est désormais effective dans ce notebook.
-Les chiffres exacts sont consignés dans `c939_run_results.json`
-(artefact gitignored à côté du notebook). Pour une comparaison
-simplifiée sur les cellules d'interprétation, voir 35 / 42 / 47 / 52 / 55.
-
-"""
-    _set_cell_source(nb.cells[61], cell61_new)
-
-    print("Writing notebook...")
-    write_nb(nb)
-    print("Done. Re-execute cells 34/41/44/51/54 via jupyter to validate.")
-
-
-if __name__ == "__main__":
-    main()
